WIP: cyclic on unstructured
